+++ src/tdx_model/data/daily_data.rs
@@ -3,15 +3,15 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DailyData {
     pub date: u32,
-    pub open: f32,
-    pub high: f32,
-    pub low: f32,
-    pub close: f32,
+    pub open: f64,
+    pub high: f64,
+    pub low: f64,
+    pub close: f64,
     // last 4 bytes are padding and should be ignored
 }
 
 impl DailyData {
-    pub fn new(date: u32, open: f32, high: f32, low: f32, close: f32) -> Self {
+    pub fn new(date: u32, open: f64, high: f64, low: f64, close: f64) -> Self {
         Self {
             date,
             open,
